+++ src/Command.php
@@ -3,6 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\CLI;
 
+use Innmind\CLI\Command\Usage;
 use Innmind\Immutable\Attempt;
 
 interface Command
@@ -15,5 +16,5 @@ interface Command
     /**
      * @psalm-mutation-free
      */
-    public function usage(): string;
+    public function usage(): Usage;
 }